requirements.txt
django~=3.2.5
gunicorn~=20.1.0
    #
mysqlclient~=2.1
    #
djangorestframework~=3.13.1
PyYAML~=6.0
uritemplate~=4.1.1
Markdown~=3.3.6
Pygments~=2.11.2
django-filter~=21.1
django-guardian~=2.4.0